Rename certificate_validator.py to CertificateValidator.py to follow project convention
